Move tested person to separate file, add marks
