+++ src/xcode/ENA/ENA/Source/Services/CoronaTestService/PCRTest.swift
@@ -6,6 +6,8 @@ import Foundation
 
 struct PCRTest: Equatable {
 
+	// MARK: - Internal
+
 	var registrationDate: Date
 	var registrationToken: String?
